The errors are not properly resolved by this change.
Because the Jenkins update center only provides dependency information
for the current version of plugins, the plugin puppet directive cannot
due dependency resolution. This means that my changes do nothing, it's
also why pipeline-step-api was listed already and still not installed
correctly.

For now we'll just need to update docs to let people know that this is
expected behavior.

There is a project that may someday be able to resolve this for us:
https://github.com/rtyler/jpm

The issue tracking the jenkins-puppet dependency problem is
https://github.com/jenkinsci/puppet-jenkins/issues/677